+++ xblock/core.py
@@ -6,11 +6,10 @@ import functools
 import inspect
 import json
 import logging
-import os
 import warnings
 from collections import OrderedDict, defaultdict
 
-import pkg_resources
+import importlib.resources
 from lxml import etree
 from webob import Response
 
@@ -157,7 +156,17 @@ class Blocklike(metaclass=_AutoNamedFieldsMetaclass):
         if "/." in uri:
             raise DisallowedFileError("Only safe file names are allowed: %r" % uri)
 
-        return pkg_resources.resource_stream(cls.__module__, os.path.join(cls.resources_dir, uri))
+        return cls._open_resource(uri)
+
+    @classmethod
+    def _open_resource(cls, uri):
+        return importlib.resources.files(
+            inspect.getmodule(cls).__package__
+        ).joinpath(
+            cls.resources_dir
+        ).joinpath(
+            uri
+        ).open('rb')
 
     @classmethod
     def json_handler(cls, func):